Fix uses of norm in regularisation docs
`norm` is part of the `LinearAlgebra` package, and is not available by default.

The output of `params` has also changed. Finally, the `activations` function has to be imported, but is broken: https://github.com/FluxML/Flux.jl/issues/702 Until this is fixed, maybe the example should be removed from the docs?