Clarify that Flux.params updates
I was confused when initally reading the overciew, because I though that `Flux.params` returned a touple of vectors. But it later becomes apparent that it actually returns another data-structure, because it updates when parameters change. I have added a single sentence to clarify this point. Edits on the wording/phrasing are very welcome.